modified:   src/components/styled-components/Button/Button.styles.js 	renamed:    src/components/ui/Navbar/Navbar.js -> src/layouts/Navbar/Navbar.js 	renamed:    src/components/ui/Navbar/nav.module.css -> src/layouts/Navbar/nav.module.css 	modified:   src/pages/index.js
