+++ src/components/styled-components/Button/Button.styles.js
@@ -18,5 +18,4 @@ export const LoginButton = styled(Button)`
   background-color: #00a8ff;
   color: #111;
   border: none;
-
 `
